Move code from AutoTLSCross to AutoTLSPlugin
Add attribution to sbt/sbt.
